Extract shared openDB into db.ts, bump IndexedDB to v2 with drafts store
